% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_estimation.R
\name{neuronActivity,matrix-method}
\alias{neuronActivity,matrix-method}
\title{Performing activity evaluation}
\usage{
\S4method{neuronActivity}{matrix}(
  x,
  gene_list,
  method = c("ssgsea", "gsva", "aucell", "zscore", "plage", "viper"),
  normalization = TRUE,
  min_target_num = 5,
  max_target_num = 500,
  norm_act = FALSE,
  scale_coe = 10,
  ncores = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{An object with gene expression data used for neuron activity evaluation,
which can be \code{matrix}, \code{SingleCellExperiment}, \code{Seurat} and \code{ExpressionSet}.}

\item{gene_list}{List. Regulons list used for activity evaluation, names of each regulons
should be Transcription Factors(TFs).}

\item{method}{Method used to calculate the active score of regulons for each sample.}

\item{normalization}{Logical, whether to normalize the activate score?}

\item{min_target_num}{Minimum number of genes in your regulons.}

\item{max_target_num}{Maximum number of genes in your regulons.}

\item{norm_act}{Whether to scale the activity score? Default: FALSE.}

\item{ncores}{Number of threads used for parallel calculation. Default: 1.}

\item{verbose}{Gives information about each calculation step. Default: TRUE.}

\item{...}{Other parameters passed to \code{\link{runViper}}, \code{\link{runAUCell}} and
\code{\link{runGSVA}}.}

\item{use_assay}{For object with \code{SummarizedExperiment} class, the use_assay
argument can be used to select the assay containing the expression data we want as input.}

\item{use_regulon}{The regulon assay used for activity estimation. Default: 1, the first one.}

\item{norm_coe}{The coefficient used for activity score normalization.}
}
\description{
Performing activity evaluation
}
